Renames routes/profiles
